+++ CHANGELOG.markdown
@@ -1,8 +1,8 @@
 3.0
 ---
-* Added `Control.Lens.Zipper`
-* Added `<<~`
-* Added `:->`, `:=>`, and `:<->` as type operator aliases for `Simple Lens`, `Simple Traversal`, and `Simple Iso`  respectively
+* Added `Control.Lens.Zipper`.
+* Added `<<~`, a version of `<~` that supports chaining assignment.
+* Added `:->`, `:=>`, and `:<->` as type operator aliases for `Simple Lens`, `Simple Traversal`, and `Simple Iso`  respectively.
 
 2.9
 ---